requirements/base.txt
Django==4.2.3
requests==2.31.0
django-filter==23.2
django-environ==0.10.0
drf-spectacular==0.26.3
psycopg2-binary==2.9.5
django-cors-headers==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2

Faker==19.2.0
factory-boy==3.3.0
pytest==7.4.0
pytest-django==4.5.2
